+++ tests/testthat/test-compare_parameters.R
@@ -81,3 +81,61 @@ if (requiet("testthat") && requiet("parameters") && requiet("insight")) {
     expect_equal(out[1], "Parameter    |    linear model (m1) |   logistic reg. (m2)")
   })
 }
+
+
+.runThisTest <- Sys.getenv("RunAllparametersTests") == "yes"
+
+if (.runThisTest && requiet("testthat") && requiet("glmmTMB") && requiet("parameters") && requiet("insight")) {
+  data("fish")
+  m0 <- glm(count ~ child + camper, data = fish, family = poisson())
+
+  m1 <- glmmTMB(
+    count ~ child + camper + (1 | persons) + (1 | ID),
+    data = fish,
+    family = poisson()
+  )
+
+  m2 <- glmmTMB(
+    count ~ child + camper + zg + (1 | ID),
+    ziformula = ~ child + (1 | persons),
+    data = fish,
+    family = truncated_poisson()
+  )
+
+  cp <- compare_parameters(m0, m1, m2, effects = "all", component = "all")
+  out <- utils::capture.output(print(cp))
+  test_that("compare_parameters, correct random effects", {
+    expect_equal(
+      cp$`Log-Mean.m2`,
+      c(
+        1.4075, -0.52925, 0.58381, 0.27885, NA, 0.13266, -0.92451,
+        1.96104, 1.07818
+      ),
+      tolerance = 1e-3
+    )
+    expect_equal(
+      cp$`Log-Mean.m1`,
+      c(0.6835, -1.67334, 0.94341, 0.26717, 1.20775, NA, NA, NA, NA),
+      tolerance = 1e-3
+    )
+    expect_equal(
+      cp$Component,
+      c(
+        "conditional", "conditional", "conditional", "conditional",
+        "conditional", "conditional", "zero_inflated", "zero_inflated",
+        "zero_inflated"
+      )
+    )
+    expect_equal(
+      cp$Effects,
+      c(
+        "fixed", "fixed", "fixed", "random", "random", "fixed", "fixed",
+        "fixed", "random"
+      )
+    )
+    expect_equal(
+      out[22],
+      "SD (Intercept: persons) |    |  1.21 ( 0.60,  2.43) |                     "
+    )
+  })
+}